# Optimize CurrentAttributes method generation